result of dynamic object validation should be less pedant
When using :goog mode (goog.object/get or /set) it makes little sense to
run the generated code for non-objects because internally goog.object/get uses
`in` operator which throws when accessing non-objects and leads to cryptic
generic errors about something wrong with "in" "key" and "obj".

I don't want to distinguish the case for :goog and :core generation,
so I decided to relax  gen-dynamic-object-access-validation a bit.
It returns false for non-objects where in would clearly fail.
Else it allows generated code to run.

related to issue #13